fix last commit

+++ wcfsetup/install/files/lib/data/language/LanguageEditor.class.php
@@ -172,13 +172,16 @@ class LanguageEditor extends DatabaseObjectEditor implements IEditableCachedObje
 		
 		// fetch deletes
 		$deletes = $xpath->query('/ns:language/ns:delete');
-		foreach ($xpath->query('child::*', $deletes) as $delete) {
-			$deleteValue = $delete->nodeValue;
-			$sql = "DELETE FROM wcf".WCF_N."_language_item where languageItem = ? AND packageID = ?";
-			$statement = WCF::getDB()->prepareStatement($sql);
-			$statement->execute(array($deleteValue, $packageID));
+		if (!empty($deletes)) {
+			foreach ($xpath->query('child::*', $deletes) as $delete) {
+				$deleteValue = $delete->nodeValue;
+				if (!empty($deleteValue)) {
+					$sql = "DELETE FROM wcf".WCF_N."_language_item where languageItem = ? AND packageID = ?";
+					$statement = WCF::getDB()->prepareStatement($sql);
+					$statement->execute(array($deleteValue, $packageID));
+				}
+			}
 		}
-		
 		
 		// fetch categories
 		$categories = $xpath->query('/ns:language/ns:category');
